Refactored python packages names.
